requirements.txt
Django==2.0.3
django-rest-swagger==2.2.0
djangorestframework==3.9.1
psycopg2==2.7.7
django-autofixture==0.12.1
